requirements.txt
google-generativeai
requests
pandas
streamlit